backend/requirements.txt
asgiref==3.11.1
django-cors-headers==4.9.0
Django==5.2.12
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
psycopg2==2.9.11
PyJWT==2.12.1
python-decouple==3.8
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2026.1
